--- site/src/components/home/Flux.js
@@ -1,126 +1,0 @@
-import React from 'react';
-import {Link} from 'gatsby';
-
-const Flux = () => {
-	return (
-		<>
-		/*
-		<section id="flux">
-
-			<h1>À chaud</h1>
-			<div id="flux-container">
-				<!-- if type=git (bon modèle pour les liens) -->
-				<article data-type="git">
-					<p className="numero">1/12</p>
-					<!---->
-					<div className="embleme">
-						<pre>░</pre>
-					</div>
-					<h1 className="type">
-						<p>Git</p>
-					</h1>
-					<hgroup>
-					<a href="#linkGit">
-						
-						<p data-level-3="action">Jacques push to</p>
-						<h1 data-level-1="title">Repo Rumba</h1>
-					</a>
-					</hgroup>
-					<div className="description">
-						<p data-level-2="label">Outils composé de 0 ou de 1, de $ ou de <<, enfin plein de machin pas si complexe...</p>
-					</div>
-					<pre className="end">░░░░░░░░░░░░░░░░░░░░░</pre>
-				</article>
-				<article data-type="git">
-					<p className="numero">2/11</p>
-					<!---->
-					<div className="embleme">
-						<pre>▒</pre>
-					</div>
-					<h1 className="type">
-						<p>Git</p>
-					</h1>
-					<hgroup>
-					<a href="#linkGit">
-						
-						<p data-level-3="action">Monique push to</p>
-						<h1 data-level-1="title">Repo Salsa</h1>
-					</a>
-					</hgroup>
-					<div className="description">
-						<p data-level-2="label">Outils composé de 0 ou de 1, de $ ou de <<, enfin plein de machin pas si complexe...</p>
-					</div>
-					<pre className="end">▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒</pre>
-				</article>
-				<article data-type="git">
-					<p className="numero">31/10</p>
-					<!---->
-					<div className="embleme">
-						<pre>▓</pre>
-					</div>
-					<h1 className="type">
-						<p>Git</p>
-					</h1>
-					<hgroup>
-					<a href="#linkGit">
-						
-						<p data-level-3="action">Rob push to</p>
-						<h1 data-level-1="title">Repo Rumba</h1>
-					</a>
-					</hgroup>
-					<div className="description">
-						<p data-level-2="label">Outils composé de 0 ou de 1, de $ ou de <<, enfin plein de machin pas si complexe...</p>
-					</div>
-					<pre className="end">▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓</pre>
-				</article>
-				<article data-type="git">
-					<p className="numero">10/10</p>
-					<!---->
-					<div className="embleme">
-						<pre>░</pre>
-					</div>
-					<h1 className="type">
-						<p>Git</p>
-					</h1>
-					<hgroup>
-					<a href="#linkGit">
-						
-						<p data-level-3="action">Jacques push to</p>
-						<h1 data-level-1="title">Repo Tcha Tcha</h1>
-					</a>
-					</hgroup>
-					<div className="description">
-						<p data-level-2="label">Outils composé de 0 ou de 1, de $ ou de <<, enfin plein de machin pas si complexe...</p>
-					</div>
-					<pre className="end">░░░░░░░░░░░░░░░░░░░░░</pre>
-				</article>
-				<article data-type="git">
-					<p className="numero">9/09</p>
-					<!---->
-					<div className="embleme">
-						<pre>▓</pre>
-					</div>
-					<h1 className="type">
-						<p>Git</p>
-					</h1>
-					<hgroup>
-					<a href="#linkGit">
-						
-						<p data-level-3="action">Jacques push to</p>
-						<h1 data-level-1="title">Repo Fox Trot</h1>
-					</a>
-					</hgroup>
-					<div className="description">
-						<p data-level-2="label">Outils composé de 0 ou de 1, de $ ou de <<, enfin plein de machin pas si complexe...</p>
-					</div>
-					<pre className="end">▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓</pre>
-				</article>			
-			</div>
-			
-		</section>
-		*/
-		</>
-  	);
-}
-
-export default Flux;
