+++ site/src/components/home/Home.js
@@ -3,14 +3,11 @@ import React from 'react';
 import {graphql} from 'gatsby';
 import {Link} from 'gatsby';
 
-import Slideshow from './home/Slideshow.js';
-import Now from './home/Now.js';
-import Agenda from './home/Agenda.js';
-import Flux from './home/Flux.js';
+import Slideshow from './fragments/Slideshow.js';
+import Now from './fragments/Now.js';
+import Agenda from './fragments/Agenda.js';
+import Flux from './fragments/Flux.js';
 
-/*import RawHtml from './RawHtml';*/
-
-//import './scss/index.scss';
 
 
 export default function Home({lang, grid, slider, rdv}) {
