+++ site/src/components/home/fragments/Agenda.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Link} from 'gatsby';
-import DateNews from '../helpers/DateNews.js';
-import TimeNews from '../helpers/TimeNews.js';
+import DateNews from '../../helpers/DateNews.js';
+import TimeNews from '../../helpers/TimeNews.js';
 import {format as formatDate, getYear, parseISO} from 'date-fns';
 
 export default function Agenda({rdv, lang}) {
