+++ site/src/components/home/fragments/Flux.js
@@ -1,0 +1,21 @@
+import React from 'react';
+import {Link} from 'gatsby';
+
+const Flux = () => {
+	return (
+		<>
+		/*
+		<section id="flux">
+
+			<h1>À chaud</h1>
+			<div id="flux-container">
+	
+			</div>
+			
+		</section>
+		*/
+		</>
+  	);
+}
+
+export default Flux;
