+++ site/src/components/pages_object/NewsDetail.js
@@ -1,18 +1,18 @@
 import React from 'react';
 import {graphql} from 'gatsby';
-import RawHtml from './helpers/RawHtml.js';
+import RawHtml from '../helpers/RawHtml.js';
 
-import Nav from './common/Nav.js';
-import ToggleLang from '.fragments/ToggleLang.js';
-import DateNews from './helpers/DateNews.js';
-import TimeNews from './helpers/TimeNews.js';
+import Nav from '../common/Nav.js';
+import ToggleLang from './fragments/ToggleLang.js';
+import DateNews from '../helpers/DateNews.js';
+import TimeNews from '../helpers/TimeNews.js';
 import {format as formatDate, getYear, parseISO} from 'date-fns';
-import {IsModel} from './helpers/helpers.js';
+import {IsModel} from '../helpers/helpers.js';
 
-import ProductionsAssociees from './pages_list/ProductionsAssociees.js';
-import ActivitesAssociees from './pages_list/ActivitesAssociees.js';
-import MembresAssocies from './pages_list/MembresAssocies.js';
-import FichiersAssocies from './pages_list/FichiersAssocies.js';
+import ProductionsAssociees from './fragments/ProductionsAssociees.js';
+import ActivitesAssociees from './fragments/ActivitesAssociees.js';
+import MembresAssocies from './fragments/MembresAssocies.js';
+import FichiersAssocies from './fragments/FichiersAssocies.js';
 
 export const queryFragment = graphql`
   fragment NewsDetail on NewsJson {
