+++ site/src/components/pages_object/fragments/FichiersAssocies.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Link, withPrefix} from 'gatsby';
-import {SECTIONS} from '../helpers/sections';
+import {SECTIONS} from '../../helpers/sections';
 
 
 const FichiersAssocies = ({lang, attachments}) => {
