+++ site/src/components/pages_object/fragments/Highlights.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import {Link} from 'gatsby';
-import {IsModel} from '../helpers/helpers.js';
-import RawHTML from '../helpers/RawHtml.js';
+import {IsModel} from '../../helpers/helpers.js';
+import RawHTML from '../../helpers/RawHtml.js';
 
-import DateNews from '../helpers/DateNews.js';
-import ProcessedImage from '../helpers/ProcessedImage.js';
+import DateNews from '../../helpers/DateNews.js';
+import ProcessedImage from '../../helpers/ProcessedImage.js';
 
 
 export default function highlights({people, lang}) {
